fix(web): use memory-optimized webpack builds
