+++ apps/web/test/next-config.test.ts
@@ -348,9 +348,11 @@ test("next.config keeps Turbopack focused on the repo root without custom worksp
   assert.equal(productionNextConfig.experimental?.cpus, HOSTED_WEB_PRODUCTION_BUILD_CPUS);
 });
 
-test("production build skips Turbopack source maps without relying on its ignored memory option", () => {
+test("production build uses the isolated memory-optimized Webpack path", () => {
   assert.equal(productionNextConfig.experimental?.turbopackMemoryLimit, undefined);
   assert.equal(productionNextConfig.experimental?.turbopackSourceMaps, false);
+  assert.equal(productionNextConfig.experimental?.webpackBuildWorker, true);
+  assert.equal(productionNextConfig.experimental?.webpackMemoryOptimizations, true);
 });
 
 test("hosted runtime issue imports avoid the runtime-state Node barrel", () => {
@@ -365,7 +367,7 @@ test("hosted runtime issue imports avoid the runtime-state Node barrel", () => {
 
 test("device connect routes use the narrow connect-config entrypoint", () => {
   for (const relativePath of [
-    "apps/web/app/(dashboard)/connect/page.tsx",
+    "apps/web/app/(dashboard)/connect/connect-page-content.tsx",
     "apps/web/app/api/connect-sources/[sourceId]/start/route.ts",
     "apps/web/app/api/internal/device-sync/connect-targets/[connectTarget]/connect-link/route.ts",
     "apps/web/app/device/connect/[claim]/route.ts",
